+++ fpdf/fpdf.py
@@ -44,6 +44,12 @@ except ImportError:
         pass
 
 
+try:
+    from endesive import signer
+    from cryptography.hazmat.primitives.serialization import pkcs12
+except ImportError:
+    signer = False
+
 from . import drawing
 from .actions import Action
 from .deprecation import WarnOnDeprecatedModuleAttributes
@@ -72,6 +78,7 @@ from .recorder import FPDFRecorder
 from .structure_tree import MarkedContent, StructureTreeBuilder
 from .svg import Percent, SVGObject
 from .syntax import DestinationXYZ
+from .syntax import build_obj_dict
 from .syntax import create_dictionary_string as pdf_dict
 from .syntax import create_list_string as pdf_list
 from .syntax import create_stream as pdf_stream
@@ -80,6 +87,7 @@ from .ttfonts import TTFontFile
 from .util import (
     enclose_in_parens,
     escape_parens,
+    format_date,
     get_scale_factor,
     object_id_for_page,
     substr,
@@ -135,6 +143,8 @@ class Annotation(NamedTuple):
     border_width: int = 0  # PDF readers support: displayed by Acrobat but not Sumatra
     name: Optional[AnnotationName] = None  # for text annotations
     ink_list: Tuple[int] = ()  # for ink annotations
+    field_type: Optional[str] = None
+    value: Optional[str] = None
 
     def serialize(self, fpdf):
         "Convert this object dictionnary to a string"
@@ -147,6 +157,12 @@ class Annotation(NamedTuple):
             f"<</Type /Annot /Subtype /{self.type}"
             f" /Rect [{rect}] /Border [0 0 {self.border_width}]"
         )
+
+        if self.field_type:
+            out += f" /FT /{self.field_type}"
+
+        if self.value:
+            out += f" /V {self.value.serialize()}"
 
         if self.flags:
             out += f" /F {sum(self.flags)}"
@@ -174,7 +190,7 @@ class Annotation(NamedTuple):
             out += f" /T ({escape_parens(self.title)})"
 
         if self.modification_time:
-            out += f" /M (D:{self.modification_time:%Y%m%d%H%M%S})"
+            out += f" /M {format_date(self.modification_time)}"
 
         if self.quad_points:
             # pylint: disable=not-an-iterable
@@ -194,6 +210,28 @@ class Annotation(NamedTuple):
             out += f" /InkList [{ink_list}]"
 
         return out + ">>"
+
+
+class Signature:
+    def __init__(self, contact_info=None, location=None, m=None, reason=None, **kwargs):
+        super().__init__(**kwargs)
+        self.type = "/Sig"
+        self.filter = "/Adobe.PPKLite"
+        self.sub_filter = "/adbe.pkcs7.detached"
+        self.contact_info = contact_info
+        "Information provided by the signer to enable a recipient to contact the signer to verify the signature"
+        self.location = location
+        "The CPU host name or physical location of the signing"
+        self.m = m
+        "The time of signing"
+        self.reason = reason
+        "The reason for the signing"
+        self.byte_range = _SIGNATURE_BYTERANGE_PLACEHOLDER
+        self.contents = "<" + _SIGNATURE_CONTENTS_PLACEHOLDER + ">"
+
+    def serialize(self):
+        obj_dict = build_obj_dict({key: getattr(self, key) for key in dir(self)})
+        return pdf_dict(obj_dict)
 
 
 class TitleStyle(NamedTuple):
@@ -373,6 +411,7 @@ class FPDF(GraphicsStateMixin):
         self._outlines_obj_id = None
         self._toc_placeholder = None  # ToCPlaceholder
         self._outline = []  # list of OutlineSection
+        self._sign_key = None
         self.section_title_styles = {}  # level -> TitleStyle
 
         # Standard fonts
@@ -435,7 +474,7 @@ class FPDF(GraphicsStateMixin):
         self.viewer_preferences = None
         self.compress = True  # Enable compression by default
         self.pdf_version = "1.3"  # Set default PDF version No.
-        self.creation_date = True
+        self.creation_date = datetime.now(timezone.utc)
 
         self._current_draw_context = None
         self._drawing_graphics_state_registry = drawing.GraphicsStateDictRegistry()
@@ -682,6 +721,10 @@ class FPDF(GraphicsStateMixin):
 
     def set_creation_date(self, date=None):
         """Sets Creation of Date time, or current time if None given."""
+        if self._sign_key:
+            raise FPDFException(
+                ".set_creation_date() must always be called before .sign*() methods"
+            )
         self.creation_date = date
 
     def set_xmp_metadata(self, xmp_metadata):
@@ -2103,7 +2146,7 @@ class FPDF(GraphicsStateMixin):
         """
         type = TextMarkupType.coerce(type).value
         if modification_time is None:
-            modification_time = datetime.now()
+            modification_time = self.creation_date
         if page is None:
             page = self.page
         x_min = min(quad_points[0::2])
@@ -3637,6 +3680,130 @@ class FPDF(GraphicsStateMixin):
                 ) from error
         return txt
 
+    def sign_pkcs12(
+        self,
+        pkcs_filepath,
+        password=None,
+        hashalgo="sha256",
+        contact_info=None,
+        location=None,
+        signing_time=None,
+        reason=None,
+        flags=(AnnotationFlag.PRINT, AnnotationFlag.LOCKED),
+    ):
+        """
+        Args:
+            pkcs_filepath (str): file path to a .pfx or .p12 PKCS12,
+                in the binary format described by RFC 7292
+            password (bytes-like): the password to use to decrypt the data.
+                `None` if the PKCS12 is not encrypted.
+            hashalgo (str): hashing algorithm used, passed to `hashlib.new`
+            contact_info (str): optional information provided by the signer to enable
+                a recipient to contact the signer to verify the signature
+            location (str): optional CPU host name or physical location of the signing
+            signing_time (datetime): optional time of signing
+            reason (str): optional signing reason
+            flags (Tuple[fpdf.enums.AnnotationFlag], Tuple[str]): optional list of flags defining annotation properties
+        """
+        with open(pkcs_filepath, "rb") as pkcs_file:
+            key, cert, extra_certs = pkcs12.load_key_and_certificates(
+                pkcs_file.read(), password
+            )
+        self.sign(
+            key=key,
+            cert=cert,
+            extra_certs=extra_certs,
+            hashalgo=hashalgo,
+            contact_info=contact_info,
+            location=location,
+            signing_time=signing_time,
+            reason=reason,
+            flags=flags,
+        )
+
+    @check_page
+    def sign(
+        self,
+        key,
+        cert,
+        extra_certs=(),
+        hashalgo="sha256",
+        contact_info=None,
+        location=None,
+        signing_time=None,
+        reason=None,
+        flags=(AnnotationFlag.PRINT, AnnotationFlag.LOCKED),
+    ):
+        """
+        Args:
+            key: certificate private key
+            cert (cryptography.x509.Certificate): certificate
+            extra_certs (list[cryptography.x509.Certificate]): list of additional PKCS12 certificates
+            hashalgo (str): hashing algorithm used, passed to `hashlib.new`
+            contact_info (str): optional information provided by the signer to enable
+                a recipient to contact the signer to verify the signature
+            location (str): optional CPU host name or physical location of the signing
+            signing_time (datetime): optional time of signing
+            reason (str): optional signing reason
+            flags (Tuple[fpdf.enums.AnnotationFlag], Tuple[str]): optional list of flags defining annotation properties
+        """
+        if not signer:
+            raise EnvironmentError(
+                "endesive.signer not available - PDF cannot be signed - Try: pip install endesive"
+            )
+
+        self._sign_key = key
+        self._sign_cert = cert
+        self._sign_extra_certs = extra_certs
+        self._sign_hashalgo = hashalgo
+
+        self.annots[self.page].append(
+            Annotation(
+                "Widget",
+                field_type="Sig",
+                x=0,
+                y=0,
+                width=0,
+                height=0,
+                flags=flags,
+                title="signature",
+                value=Signature(
+                    contact_info=contact_info,
+                    location=location,
+                    m=format_date(signing_time) if signing_time else None,
+                    reason=reason,
+                ),
+            )
+        )
+
+    def _sign_content(self):
+        """
+        Perform PDF signing based on ._sign_* properties
+        and the content of the .buffer, performing substitutions in it.
+        """
+        contents_placeholder = _SIGNATURE_CONTENTS_PLACEHOLDER.encode("latin1")
+        start_index = self.buffer.find(contents_placeholder)
+        end_index = start_index + len(contents_placeholder)
+        range = (0, start_index - 1, end_index + 1, len(self.buffer) - end_index - 1)
+        br_placeholder = _SIGNATURE_BYTERANGE_PLACEHOLDER.encode()
+        byte_range = b"[%010d %010d %010d %010d]" % range
+        self.buffer = self.buffer.replace(br_placeholder, byte_range, 1)
+        # We compute the ByteRange hash, of everything before & after the placeholder:
+        hash = hashlib.new(self._sign_hashalgo)
+        hash.update(self.buffer[: range[1]])  # before
+        hash.update(self.buffer[range[2] :])  # after
+        contents = signer.sign(
+            datau=None,
+            key=self._sign_key,
+            cert=self._sign_cert,
+            othercerts=self._sign_extra_certs,
+            hashalgo=self._sign_hashalgo,
+            attrs=True,
+            signed_value=hash.digest(),
+        )
+        contents = _pkcs11_aligned(contents).encode("latin1")
+        self.buffer = self.buffer.replace(contents_placeholder, contents, 1)
+
     def _putpages(self):
         nb = self.pages_count  # total number of pages
         if self.str_alias_nb_pages:
@@ -4253,14 +4420,9 @@ class FPDF(GraphicsStateMixin):
             "/Producer": enclose_in_parens(getattr(self, "producer", None)),
         }
 
-        if self.creation_date is True:
-            # => no date has been specified, we use the current time by default:
-            self.creation_date = datetime.now(timezone.utc)
         if self.creation_date:
             try:
-                info_d["/CreationDate"] = enclose_in_parens(
-                    f"D:{self.creation_date:%Y%m%d%H%M%SZ%H'%M'}"
-                )
+                info_d["/CreationDate"] = format_date(self.creation_date, with_tz=True)
             except Exception as error:
                 raise FPDFException(
                     f"Could not format date: {self.creation_date}"
@@ -4388,6 +4550,8 @@ class FPDF(GraphicsStateMixin):
             self._out("startxref")
             self._out(o)
         self._out("%%EOF")
+        if self._sign_key:
+            self._sign_content()
         self.state = DocumentState.CLOSED
 
     def _beginpage(
@@ -4852,6 +5016,15 @@ def _sizeof_fmt(num, suffix="B"):
 
 def _is_svg(bytes):
     return bytes.startswith(b"<?xml ") or bytes.startswith(b"<svg ")
+
+
+def _pkcs11_aligned(data):
+    data = "".join(f"{i:02x}" for i in data)
+    return data + "0" * (0x4000 - len(data))
+
+
+_SIGNATURE_BYTERANGE_PLACEHOLDER = "[0000000000 0000000000 0000000000 0000000000]"
+_SIGNATURE_CONTENTS_PLACEHOLDER = _pkcs11_aligned((0,))
 
 
 sys.modules[__name__].__class__ = WarnOnDeprecatedModuleAttributes