+++ test/metadata/test_file_id.py
@@ -9,9 +9,7 @@ HERE = Path(__file__).resolve().parent
 
 def test_default_file_id():
     pdf = FPDF()
-    pdf.creation_date = (
-        EPOCH  # this property always receive a default value when calling .output()
-    )
+    pdf.creation_date = EPOCH
     assert (
         pdf.file_id()
         == "<A29419ACCA57A4335F99B957670E2C04><A29419ACCA57A4335F99B957670E2C04>"